Make GetToken return int64 instead of uint32

+++ bindings/dm.go
@@ -334,8 +334,8 @@ func (dmc *DMClient) GetPublicKey() []byte {
 }
 
 // GetToken returns the DM token of this client.
-func (dmc *DMClient) GetToken() uint32 {
-	return dmc.api.GetToken()
+func (dmc *DMClient) GetToken() int64 {
+	return int64(dmc.api.GetToken())
 }
 
 // GetIdentity returns the public identity associated with this client.
